Move CommitChart
