+++ src/pages/commits/ActiveCommits/CommitChart/CommitChart.stories.tsx
@@ -1,11 +1,11 @@
 import styled from "@emotion/styled";
 import { StoryObj } from "@storybook/react";
-import { ChartTypes, Commits } from "types/commits";
 import {
   findMaxGroupedTaskStats,
   getAllTaskStatsGroupedByColor,
-} from "../utils";
-import { CommitChart } from "./CommitChart";
+} from "pages/commits/utils";
+import { ChartTypes, Commits } from "types/commits";
+import { CommitChart } from ".";
 
 export default {
   title: "Pages/Commits/Charts",
